refactor: properly updated deps & names for long/short yes/no dialogs
