+++ lib/dialogs/predict_no/bloc/no_button_state.dart
@@ -1,17 +1,15 @@
-part of 'long_button_bloc.dart';
+part of 'no_button_bloc.dart';
 
-class YesButtonState extends Equatable {
-  const YesButtonState({
+class NoButtonState extends Equatable {
+  const NoButtonState({
     this.status = BlocStatus.initial,
     this.errorMessage = '',
-    this.swapInfo = SwapInfo.empty,
     this.aptBuyInfo = AptBuyInfo.empty,
     this.aptSellInfo = AptSellInfo.empty,
   });
 
   final BlocStatus status;
   final String errorMessage;
-  final SwapInfo swapInfo;
   final AptBuyInfo aptBuyInfo;
   final AptSellInfo aptSellInfo;
 
@@ -20,21 +18,18 @@ class YesButtonState extends Equatable {
         status,
         aptBuyInfo,
         aptSellInfo,
-        swapInfo,
         errorMessage,
       ];
 
-  YesButtonState copyWith({
+  NoButtonState copyWith({
     BlocStatus? status,
     String? errorMessage,
-    SwapInfo? swapInfo,
     AptBuyInfo? aptBuyInfo,
     AptSellInfo? aptSellInfo,
   }) {
-    return YesButtonState(
+    return NoButtonState(
       status: status ?? this.status,
       errorMessage: errorMessage ?? this.errorMessage,
-      swapInfo: swapInfo ?? this.swapInfo,
       aptBuyInfo: aptBuyInfo ?? this.aptBuyInfo,
       aptSellInfo: aptSellInfo ?? this.aptSellInfo,
     );
